refactor(bootstrap): inline package root tests
